refactor(api): split shutdownDeps into runtime_manager_shutdown.go
